requirements/marllib.txt
# MARLlib runtime dependencies (requires Python 3.8 or 3.9)
# Recommended environment setup:
#   conda create -n marllib python=3.9
#   conda activate marllib
#   pip install -r requirements/marllib.txt

ray[rllib]==1.8.0
ray[tune]==1.8.0
ray==1.8.0
numpy==1.20.3
scipy==1.7.3
torch==1.9.0
pettingzoo[mpe]==1.12.0
supersuit==3.2.0
overcooked-ai==1.2.0
gym==0.20.0
pyyaml==5.4.1
pandas==1.3.5
matplotlib==3.4.3
seaborn==0.11.2
tabulate==0.9.0
cloudpickle==1.6.0
wandb==0.15.12

# Optional: MA-MuJoCo continuous control tasks
mamujoco @ git+https://github.com/Replicable-MARL/MARLlib-mamujoco@b7a1c47c2f066fa251cc403642c7e09622cb3f27
